style: 일기 뷰 페이지 스타일 수정
