+++ src/components/CreatedDiary/CreatedDiary.style.js
@@ -4,17 +4,12 @@ export const CreatedDiaryWrapper = styled.div`
     width: 32rem;
     margin-bottom: 2.5rem;
 `
-export const HeaderWrapper = styled.div`
-    ${({ theme: { mixin } }) => mixin.flexBox({})};
-    margin-top: 3.2rem;
-    justify-content: space-between;
-`
+
 export const CreatComponentWrapper = styled.div`
     width: 100%;
     ${({ theme }) => theme.colors.gradient.gradient_pink};
 `
-export const BtnBackWrapper = styled.div`
-`
+
 export const CreatedDiaryComponentWrapper = styled.div`
     ${({ theme: { mixin } }) => mixin.flexCenter({})};
     gap: 1.1rem;
@@ -75,16 +70,7 @@ export const DiaryText = styled.p`
     margin-bottom: 4rem;    
     white-space: pre-line;
 `
-export const ExtraBtnWrapper = styled.div`
-    ${({ theme: { mixin } }) => mixin.flexCenter({direction:'row'})};
-    background-color: transparent;
-`
-export const MenuWrapper = styled.div`
-    width: 10rem;
-    height: 8rem;
-    background-color: white;
 
-`
 export const EditBtn = styled.button`
     ${({ theme }) => theme.fonts.body_02};
 `
